+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.11.0] / 2026-02-11
 ### Updates
 - Remove `Nuke.Common.Tools.MSBuild` to use `Nuke.Common.Tools.DotNet` for build and rebuild. (Fix: #85)
+- Add `GetConfigurations` in `net10` to get all configurations in the solution/project.
+- Disable warning in `net8.0`.
 
 ## [1.10.0] / 2026-01-12 - 2026-01-21
 ### Features